Add windows/artifacts/binaries target and associated tests Separate handlers for artifacts/binaries and zip into different files and regroup common windows functionality under windows/common.go
