+++ frontend/windows/common.go
@@ -1,0 +1,113 @@
+package windows
+
+import (
+	"bytes"
+	"context"
+	"fmt"
+	"path/filepath"
+	"strings"
+
+	"github.com/Azure/dalec"
+	"github.com/Azure/dalec/frontend"
+	"github.com/moby/buildkit/client/llb"
+	gwclient "github.com/moby/buildkit/frontend/gateway/client"
+	"github.com/pkg/errors"
+)
+
+const (
+	workerImgRef    = "mcr.microsoft.com/mirror/docker/library/ubuntu:jammy"
+	outputDir       = "/tmp/output"
+	buildScriptName = "_build.sh"
+	aptCachePrefix  = "jammy-windowscross"
+)
+
+func buildBinaries(ctx context.Context, spec *dalec.Spec, worker llb.State, client gwclient.Client, sOpt dalec.SourceOpts, targetKey string) (llb.State, error) {
+	worker = worker.With(installBuildDeps(spec.GetBuildDeps(targetKey)))
+
+	sources, err := specToSourcesLLB(worker, spec, sOpt)
+	if err != nil {
+		return llb.Scratch(), errors.Wrap(err, "could not generate sources")
+	}
+
+	patched := dalec.PatchSources(worker, spec, sources)
+	buildScript := createBuildScript(spec)
+	script := generateInvocationScript(spec.Artifacts.Binaries)
+
+	pg := dalec.ProgressGroup("Build binaries")
+
+	st := worker.Run(
+		shArgs(script.String()),
+		llb.Dir("/build"),
+		withSourcesMounted("/build", patched, spec.Sources),
+		llb.AddMount("/tmp/scripts", buildScript),
+		llb.Network(llb.NetModeNone),
+		pg,
+	).AddMount(outputDir, llb.Scratch())
+
+	if signer, ok := spec.GetSigner(targetKey); ok {
+		signed, err := frontend.ForwardToSigner(ctx, client, signer, st)
+		if err != nil {
+			return llb.Scratch(), err
+		}
+
+		st = signed
+	}
+
+	return st, nil
+}
+
+func generateInvocationScript(binaryArtifacts map[string]dalec.ArtifactConfig) *strings.Builder {
+	script := &strings.Builder{}
+	fmt.Fprintln(script, "#!/usr/bin/env sh")
+	fmt.Fprintln(script, "set -ex")
+	fmt.Fprintf(script, "/tmp/scripts/%s\n", buildScriptName)
+	for path, bin := range binaryArtifacts {
+		fmt.Fprintf(script, "mv '%s' '%s'\n", path, filepath.Join(outputDir, bin.ResolveName(path)))
+	}
+	return script
+}
+
+func workerImg(sOpt dalec.SourceOpts, opts ...llb.ConstraintsOpt) llb.State {
+	// TODO: support named context override... also this should possibly be its own image, maybe?
+	return llb.Image(workerImgRef, llb.WithMetaResolver(sOpt.Resolver), dalec.WithConstraints(opts...)).
+		Run(
+			shArgs("apt-get update && apt-get install -y build-essential binutils-mingw-w64 g++-mingw-w64-x86-64 gcc git make pkg-config quilt zip"),
+			dalec.WithMountedAptCache(aptCachePrefix),
+		).Root()
+}
+
+func shArgs(cmd string) llb.RunOption {
+	return llb.Args([]string{"sh", "-c", cmd})
+}
+
+func createBuildScript(spec *dalec.Spec) llb.State {
+	buf := bytes.NewBuffer(nil)
+
+	fmt.Fprintln(buf, "#!/usr/bin/env sh")
+	fmt.Fprintln(buf, "set -x")
+
+	if spec.HasGomods() {
+		fmt.Fprintln(buf, "export GOMODCACHE=\"$(pwd)/"+gomodsName+"\"")
+	}
+
+	for i, step := range spec.Build.Steps {
+		fmt.Fprintln(buf, "(")
+
+		for k, v := range step.Env {
+			fmt.Fprintf(buf, "export %s=\"%s\"", k, v)
+		}
+
+		fmt.Fprintln(buf, step.Command)
+		fmt.Fprintf(buf, ")")
+
+		if i < len(spec.Build.Steps)-1 {
+			fmt.Fprintln(buf, " && \\")
+			continue
+		}
+
+		fmt.Fprintf(buf, "\n")
+	}
+
+	return llb.Scratch().
+		File(llb.Mkfile(buildScriptName, 0o770, buf.Bytes()))
+}
